+++ frontend/src/pages/ManagerLogin/LoginForm.styles.ts
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+export const Form = styled.form`
+  margin: 3.75rem 0 1rem;
+
+  label {
+    margin-bottom: 3rem;
+  }
+`;
